bf update

+++ Game/src/Sprites/Sheep.java
@@ -3,7 +3,7 @@ package Sprites;
 import javafx.scene.image.ImageView;
 
 public class Sheep extends Sprite{
-    public Sheep( int row, int column, int hp, ImageView imageView, int width, int heigth) {
+    public Sheep( int row, int column, int hp, ImageView imageView, double width, double heigth) {
         super(row, column,hp, width, heigth);
     }
