+++ Game/src/Sprites/Sprite.java
@@ -10,18 +10,12 @@ public class Sprite {
     private int row;
     private int column;
     private int hp;
-    private int width;
-    private int height;
-
-//    ImageView knightBlue = new ImageView(Settings.KNIGHT_BLUE);
-//    ImageView knightPurple = new ImageView(Settings.KNIGHT_PURPLE);
-//    ImageView knightRed = new ImageView(Settings.KNIGHT_RED);
-//    ImageView knightYellow = new ImageView(Settings.KNIGHT_YELLOW);
+    private double width;
+    private double height;
 
     private Type type;
 
-
-    public Sprite(int hp, int row, int column, int width, int heigth) {
+    public Sprite(int hp, int row, int column, double width, double heigth) {
         this.column = column;
         this.row = row;
         this.hp = hp;
